Addressed comments, and added base URL prepending to links in description
