+++ streaming/src/main/scala/org/apache/spark/streaming/StreamingContext.scala
@@ -198,7 +198,9 @@ class StreamingContext private[streaming] (
 
   private var state: StreamingContextState = INITIALIZED
 
-  private[streaming] val startSite = new AtomicReference[CallSite](null)
+  private val startSite = new AtomicReference[CallSite](null)
+
+  private[streaming] def getStartSite(): CallSite = startSite.get()
 
   private var shutdownHookRef: AnyRef = _
 
@@ -735,7 +737,7 @@ object StreamingContext extends Logging {
         throw new IllegalStateException(
           "Only one StreamingContext may be started in this JVM. " +
             "Currently running StreamingContext was started at" +
-            activeContext.get.startSite.get.longForm)
+            activeContext.get.getStartSite().longForm)
       }
     }
   }
